requirements.txt
autopep8==2.0.1
couchbase==4.1.1
kafka-python==2.0.2
pycodestyle==2.10.0
tomli==2.0.1
